requirements.txt
asgiref==3.8.1
Django==4.2.19
python-dotenv==1.0.1
setuptools==56.0.0
sqlparse==0.5.3
tzdata==2025.1
wheel==0.45.1
boto3==1.36.17
django-cors-headers==4.4.0
django-rest-auth==0.9.5
django-storages==1.14.4
djangorestframework_simplejwt==5.3.1
gunicorn==23.0.0
mysqlclient==2.2.7
PyMySQL==1.1.1 
django-environ==0.11.2
requests==2.32.3
boto3==1.36.17
python-dotenv==1.0.1